Upgrade docker version images
