+++ sql-db/hibernate/src/test/java/io/quarkus/qe/hibernate/ProdHibernateIT.java
@@ -11,7 +11,7 @@ public class ProdHibernateIT extends BaseHibernateIT {
 
     static final int POSTGRESQL_PORT = 5432;
 
-    @Container(image = "${postgresql.10.image}", port = POSTGRESQL_PORT, expectedLog = "listening on IPv4 address")
+    @Container(image = "${postgresql.13.image}", port = POSTGRESQL_PORT, expectedLog = "listening on IPv4 address")
     static final PostgresqlService database = new PostgresqlService();
 
     @QuarkusApplication
