+++ sql-db/sql-app/src/test/java/io/quarkus/ts/sqldb/sqlapp/MySqlDatabaseIT.java
@@ -11,7 +11,7 @@ public class MySqlDatabaseIT extends AbstractSqlDatabaseIT {
 
     static final int MYSQL_PORT = 3306;
 
-    @Container(image = "${mysql.80.image}", port = MYSQL_PORT, expectedLog = "Only MySQL server logs after this point")
+    @Container(image = "${mysql.57.image}", port = MYSQL_PORT, expectedLog = "ready for connections")
     static MySqlService database = new MySqlService();
 
     @QuarkusApplication
