+++ sql-db/sql-app/src/test/java/io/quarkus/ts/sqldb/sqlapp/OpenShiftPostgresql12DatabaseIT.java
@@ -2,9 +2,24 @@ package io.quarkus.ts.sqldb.sqlapp;
 
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
+import io.quarkus.test.bootstrap.PostgresqlService;
+import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.OpenShiftScenario;
+import io.quarkus.test.services.Container;
+import io.quarkus.test.services.QuarkusApplication;
 
 @OpenShiftScenario
 @EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
-public class OpenShiftPostgresql12DatabaseIT extends Postgresql12DatabaseIT {
+public class OpenShiftPostgresql12DatabaseIT extends AbstractSqlDatabaseIT {
+
+    static final int POSTGRESQL_PORT = 5432;
+
+    @Container(image = "${postgresql.12.image}", port = POSTGRESQL_PORT, expectedLog = "listening on IPv4 address")
+    static PostgresqlService database = new PostgresqlService();
+
+    @QuarkusApplication
+    static RestService app = new RestService().withProperties("postgresql.properties")
+            .withProperty("quarkus.datasource.username", database.getUser())
+            .withProperty("quarkus.datasource.password", database.getPassword())
+            .withProperty("quarkus.datasource.jdbc.url", database::getJdbcUrl);
 }
