--- sql-db/sql-app/src/test/java/io/quarkus/ts/sqldb/sqlapp/Postgresql10DatabaseIT.java
@@ -1,22 +1,0 @@
-package io.quarkus.ts.sqldb.sqlapp;
-
-import io.quarkus.test.bootstrap.PostgresqlService;
-import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.QuarkusScenario;
-import io.quarkus.test.services.Container;
-import io.quarkus.test.services.QuarkusApplication;
-
-@QuarkusScenario
-public class Postgresql10DatabaseIT extends AbstractSqlDatabaseIT {
-
-    static final int POSTGRESQL_PORT = 5432;
-
-    @Container(image = "${postgresql.10.image}", port = POSTGRESQL_PORT, expectedLog = "listening on IPv4 address")
-    static PostgresqlService database = new PostgresqlService();
-
-    @QuarkusApplication
-    static RestService app = new RestService().withProperties("postgresql.properties")
-            .withProperty("quarkus.datasource.username", database.getUser())
-            .withProperty("quarkus.datasource.password", database.getPassword())
-            .withProperty("quarkus.datasource.jdbc.url", database::getJdbcUrl);
-}
