+++ sql-db/sql-app/src/test/java/io/quarkus/ts/sqldb/sqlapp/Postgresql13DatabaseIT.java
@@ -1,21 +1,21 @@
 package io.quarkus.ts.sqldb.sqlapp;
 
-import io.quarkus.test.bootstrap.MariaDbService;
+import io.quarkus.test.bootstrap.PostgresqlService;
 import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.QuarkusScenario;
 import io.quarkus.test.services.Container;
 import io.quarkus.test.services.QuarkusApplication;
 
 @QuarkusScenario
-public class MariaDB102DatabaseIT extends AbstractSqlDatabaseIT {
+public class Postgresql13DatabaseIT extends AbstractSqlDatabaseIT {
 
-    static final int MARIADB_PORT = 3306;
+    static final int POSTGRESQL_PORT = 5432;
 
-    @Container(image = "${mariadb.102.image}", port = MARIADB_PORT, expectedLog = "Only MySQL server logs after this point")
-    static MariaDbService database = new MariaDbService();
+    @Container(image = "${postgresql.13.image}", port = POSTGRESQL_PORT, expectedLog = "listening on IPv4 address")
+    static PostgresqlService database = new PostgresqlService();
 
     @QuarkusApplication
-    static RestService app = new RestService().withProperties("mariadb_app.properties")
+    static RestService app = new RestService().withProperties("postgresql.properties")
             .withProperty("quarkus.datasource.username", database.getUser())
             .withProperty("quarkus.datasource.password", database.getPassword())
             .withProperty("quarkus.datasource.jdbc.url", database::getJdbcUrl);
