+++ sql-db/vertx-sql/src/test/java/io/quarkus/qe/vertx/sql/handlers/PostgresqlHandlerIT.java
@@ -10,7 +10,7 @@ import io.quarkus.test.services.QuarkusApplication;
 public class PostgresqlHandlerIT extends CommonTestCases {
     private static final String POSTGRESQL_DATABASE = "amadeus";
 
-    @Container(image = "${postgresql.10.image}", port = 5432, expectedLog = "listening on IPv4 address")
+    @Container(image = "${postgresql.13.image}", port = 5432, expectedLog = "listening on IPv4 address")
     static PostgresqlService postgres = new PostgresqlService()
             .with("test", "test", POSTGRESQL_DATABASE);
